Windows build working

+++ packager/media/codecs/aac_audio_specific_config.cc
@@ -142,9 +142,9 @@ bool AACAudioSpecificConfig::ConvertToADTS(std::vector<uint8_t>* buffer) const {
   adts[1] = 0xf1;
   adts[2] = ((audio_object_type_ - 1) << 6) + (frequency_index_ << 2) +
       (channel_config_ >> 2);
-  adts[3] = ((channel_config_ & 0x3) << 6) + (size >> 11);
-  adts[4] = (size & 0x7ff) >> 3;
-  adts[5] = ((size & 7) << 5) + 0x1f;
+  adts[3] = ((channel_config_ & 0x3) << 6) + static_cast<uint8_t>(size >> 11);
+  adts[4] = static_cast<uint8_t>((size & 0x7ff) >> 3);
+  adts[5] = static_cast<uint8_t>(((size & 7) << 5) + 0x1f);
   adts[6] = 0xfc;
 
   return true;
